Implemented refactor on the second main function of HelpfulFormatter, GetBasicTokenParser. No real logic change was done (I had already tried to make it as fast as I could), but gpt helped make the code actually readable.

+++ BLPPCounter/Utils/Misc Classes/FormatWrapper.cs
@@ -214,14 +214,14 @@ public class FormatWrapper
         if (key < SPLIT)
         {
             int idx = key - minTokenLow;
-            return (lookupLow.Length > 0 && (uint)idx < (uint)lookupLow.Length)
+            return (lookupLow.Length > 0 && idx < lookupLow.Length)
                 ? lookupLow[idx]
                 : -1;
         }
         else
         {
             int idx = key - minTokenHigh;
-            return (lookupHigh.Length > 0 && (uint)idx < (uint)lookupHigh.Length)
+            return (lookupHigh.Length > 0 && idx < lookupHigh.Length)
                 ? lookupHigh[idx]
                 : -1;
         }
